refactor(dao): remove legacy dao package
- replace with new daos using Spring JDBC
